requirements.txt
langchain
langchain-google-genai
langchain-community
pinecone
python-dotenv
pypdf
tqdm
